service/requirements.txt
numpy==1.24.2
pandas==2.0.0
scikit-learn==1.3.1
phe==1.5.0
pyzmq==22.3.0
